requirements.txt
torch>=2.2.0
torchvision>=0.17.0
monai>=1.3.0
numpy>=1.26.0
pandas>=2.1.0
opencv-python
pillow
scikit-learn
tqdm
matplotlib
albumentations